refactor(bin): move entrypoints to a proper folder
