+++ Dockerfile
@@ -7,20 +7,6 @@ FROM node:$NODE_VERSION as build
 ARG APP_DIR
 WORKDIR $APP_DIR
 
-COPY package.json yarn.lock ./
-RUN yarn install
-COPY . .
-RUN yarn build
-
-
-# # Build the app
-# FROM node:$NODE_VERSION as release
-
-# ARG APP_DIR
-# WORKDIR $APP_DIR
-
-# COPY --from=build $APP_DIR/dist .
-
-# RUN npm install -g pm2
-
-# CMD ["pm2-runtime", "bot.js"]
+RUN npm install pm2 -g
+COPY dist/ .
+CMD ["pm2-runtime", "bin/bot.js"]
