+++ bin/bot.ts
@@ -1,10 +1,10 @@
 import dotenv from 'dotenv'
 dotenv.config()
 
-import Sentry from './src/config/sentry'
-import messageHandler from './src/handlers/message'
-import readyHandler from './src/handlers/ready'
-import { client } from './src/services/discord'
+import Sentry from '../src/config/sentry'
+import messageHandler from '../src/handlers/message'
+import readyHandler from '../src/handlers/ready'
+import { client } from '../src/services/discord'
 
 client.on('ready', readyHandler)
 client.on('message', messageHandler)
