--- docker-compose.yml
@@ -1,7 +1,0 @@
-version: '3'
-
-services: 
-  bot:
-    build: .
-    command: yarn build
-
